+++ Nos/Models/NostrConstants.swift
@@ -1,0 +1,15 @@
+//
+//  NostrConstants.swift
+//  Nos
+//
+//  Created by Matthew Lorentz on 2/21/23.
+//
+
+import Foundation
+
+/// A collection of constants used in Nostr.
+enum Nostr {
+    // https://github.com/nostr-protocol/nips/blob/master/19.md
+    static let privateKeyPrefix = "nsec"
+    static let publicKeyPrefix = "npub"
+}