+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added ability to delete relays on the Relays screen.
 - Fix events not being signed correctly with a key generated during onboarding.
+- Verify signatures on events.
 - Request only events from user's Follows
 
 ## [0.1 (3)] - 2023-02-20Z